// apps/odonto-front/src/generated/ts/CreateTreatmentPlanDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { TreatmentPlanItemDto } from './TreatmentPlanItemDto.ts';

export const createTreatmentPlanDtoStatusEnum = {
  DRAFT: 'DRAFT',
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;

export type CreateTreatmentPlanDtoStatusEnumKey =
  (typeof createTreatmentPlanDtoStatusEnum)[keyof typeof createTreatmentPlanDtoStatusEnum];

export type CreateTreatmentPlanDto = {
  /**
   * @type number
   */
  patientId: number;
  /**
   * @type number
   */
  dentistId: number;
  /**
   * @type number | undefined
   */
  discount?: number;
  /**
   * @type string | undefined
   */
  status?: CreateTreatmentPlanDtoStatusEnumKey;
  /**
   * @type string | undefined
   */
  title?: string;
  /**
   * @type string | undefined
   */
  notes?: string;
  /**
   * @type array
   */
  items: TreatmentPlanItemDto[];
};