// apps/odonto-front/src/generated/ts/TreatmentPlanItemDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const treatmentPlanItemDtoStatusEnum = {
  PLANNED: "PLANNED",
  IN_PROGRESS: "IN_PROGRESS",
  COMPLETED: "COMPLETED",
  CANCELED: "CANCELED",
  CANCELLED: "CANCELLED",
} as const;

export type TreatmentPlanItemDtoStatusEnumKey =
  (typeof treatmentPlanItemDtoStatusEnum)[keyof typeof treatmentPlanItemDtoStatusEnum];

export type TreatmentPlanItemDto = {
  /**
   * @type number | undefined
   */
  id?: number;
  /**
   * @type string
   */
  description: string;
  /**
   * @type number
   */
  value: number;
  /**
   * @type number | undefined
   */
  toothNumber?: number;
  /**
   * @type object | undefined
   */
  surface?: object;
  /**
   * @type string | undefined
   */
  status?: TreatmentPlanItemDtoStatusEnumKey;
};