// apps/odonto-front/src/generated/ts/UpdateTreatmentPlanDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { TreatmentPlanItemDto } from './TreatmentPlanItemDto.ts';

export const updateTreatmentPlanDtoStatusEnum = {
  DRAFT: 'DRAFT',
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;

export type UpdateTreatmentPlanDtoStatusEnumKey =
  (typeof updateTreatmentPlanDtoStatusEnum)[keyof typeof updateTreatmentPlanDtoStatusEnum];

export type UpdateTreatmentPlanDto = {
  /**
   * @type number | undefined
   */
  patientId?: number;
  /**
   * @type number | undefined
   */
  dentistId?: number;
  /**
   * @type number | undefined
   */
  discount?: number;
  /**
   * @type string | undefined
   */
  status?: UpdateTreatmentPlanDtoStatusEnumKey;
  /**
   * @type string | undefined
   */
  title?: string;
  /**
   * @type string | undefined
   */
  notes?: string;
  /**
   * @type array | undefined
   */
  items?: TreatmentPlanItemDto[];
};